Fix ref path is not correct in document snapshot
The current implementation simply takes the collection path (for example, `/accounts/{accountId}/transactions` and the ID, to end up with a ref path of `/accounts/{accountId}/transactions/xxxxxxx`). This is correct for the trigger path, but it is **not** correct for the actual reference. This commit fixes these issues and provides the correct path.